tests/lookups: add Stella + xmile models - 2 version of Stella model
From reading the XMILE spec (section 3.3.2 - Function Calls [1]), it
seems like the Vensim-style of one aux that just has the graphical
function definition, and another aux to compute the result of looking
up something in that table (this way you can define a table once, but
do several lookups on it without redefining the function) should be
legal.  But Stella 10.0.6 complains "Syntax error: Unexpected text".

1 - http://docs.oasis-open.org/xmile/xmile/v1.0/cos01/xmile-v1.0-cos01.html#_Toc426543507